refactor: 로직 책임 변경 및 출석 종료 시 출석정보 변경 테스트 작성
