+++ src/test/kotlin/nexters/admin/service/attendance/QrCodeServiceTest.kt
@@ -29,15 +29,4 @@ class QrCodeServiceTest(
         actual.type shouldBe AttendanceStatus.ATTENDED
         actual.isExpired() shouldBe false
     }
-
-    @Test
-    fun `출석 체크 종료 이후에 유효한 QR 코드 조회 시도시 예외 발생`() {
-        qrCodeService.initializeCodes(1L, AttendanceStatus.ATTENDED.name)
-
-        qrCodeService.endAttendance()
-        shouldThrow<NotFoundException> {
-            qrCodeService.getCurrentQrCode()
-        }
-    }
-    // TODO: 출석 체크 종료시, PENDING 상태는 전부 무단 결석으로 변하는지 검증하는 테스트 코드 필요
 }
